Expose recently introduced map_array function in skimage.util (#4646)
* move map_array to util

* moves tests

* Apply suggestions from code review
