+++ skimage/util/__init__.py
@@ -14,6 +14,7 @@ from ._regular_grid import regular_grid, regular_seeds
 from .unique import unique_rows
 from ._invert import invert
 from ._montage import montage
+from ._map_array import map_array
 
 
 @functools.wraps(np.pad)
@@ -37,6 +38,7 @@ __all__ = ['img_as_float32',
            'pad',
            'crop',
            'compare_images',
+           'map_array',
            'montage',
            'random_noise',
            'regular_grid',
